backend/requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
django-filter
psycopg2-binary
django-cors-headers
